+++ linopy/solvers.py
@@ -24,7 +24,16 @@ from linopy.constants import (
     TerminationCondition,
 )
 
-QUADRATIC_SOLVERS = ["gurobi", "xpress", "cplex", "highs", "scip"]
+QUADRATIC_SOLVERS = [
+    "gurobi",
+    "xpress",
+    "cplex",
+    "highs",
+    "scip",
+    "mosek",
+    "copt",
+    "mindopt",
+]
 
 available_solvers = []
 
@@ -386,6 +395,12 @@ def run_highs(
 
     if warmstart_fn:
         logger.warning("Warmstart not available with HiGHS solver. Ignore argument.")
+
+    if solver_options.get("solver") in ["simplex", "ipm"] and model.type == "QP":
+        logger.warning(
+            "The HiGHS solver ignores quadratic terms if the solver is set to 'simplex' or 'ipm'. "
+            "Drop the solver option or use 'choose' to enable quadratic terms."
+        )
 
     if io_api is None or io_api in ["lp", "mps"]:
         model.to_file(problem_fn)
@@ -855,7 +870,7 @@ def run_mosek(
         "solsta.optimal": "optimal",
         "solsta.integer_optimal": "optimal",
         "solsta.prim_infeas_cer": "infeasible",
-        "solsta.dual_infeas_cer": "infeasible",
+        "solsta.dual_infeas_cer": "infeasible_or_unbounded",
     }
 
     if io_api is not None and io_api not in ["lp", "mps"]:
@@ -927,7 +942,7 @@ def run_mosek(
                     dual = {m.getconname(i): dual[i] for i in range(m.getnumcon())}
                     dual = pd.Series(dual, dtype=float)
                     dual = set_int_index(dual)
-                except mosek.Error:
+                except (mosek.Error, AttributeError):
                     logger.warning("Dual values of MILP couldn't be parsed")
                     dual = pd.Series(dtype=float)
 
@@ -1010,13 +1025,13 @@ def run_copt(
         except coptpy.CoptError as err:
             logger.info("No model basis stored. Raised error: %s", err)
 
-    condition = m.LpStatus if model.type == "LP" else m.MipStatus
+    condition = m.LpStatus if model.type in ["LP", "QP"] else m.MipStatus
     termination_condition = CONDITION_MAP.get(condition, condition)
     status = Status.from_termination_condition(termination_condition)
     status.legacy_status = condition
 
     def get_solver_solution() -> Solution:
-        objective = m.LpObjval if model.type == "LP" else m.BestObj
+        objective = m.LpObjval if model.type in ["LP", "QP"] else m.BestObj
 
         sol = pd.Series({v.name: v.x for v in m.getVars()}, dtype=float)
         sol = set_int_index(sol)
@@ -1024,7 +1039,7 @@ def run_copt(
         try:
             dual = pd.Series({v.name: v.pi for v in m.getConstrs()}, dtype=float)
             dual = set_int_index(dual)
-        except coptpy.CoptError:
+        except (coptpy.CoptError, AttributeError):
             logger.warning("Dual values of MILP couldn't be parsed")
             dual = pd.Series(dtype=float)
 
@@ -1072,6 +1087,10 @@ def run_mindopt(
         raise ValueError(
             "Keyword argument `io_api` has to be one of `lp`, `mps` or None"
         )
+    if (io_api == "lp" or str(problem_fn).endswith(".lp")) and model.type == "QP":
+        raise ValueError(
+            "MindOpt does not support QP problems in LP format. Use `io_api='mps'` instead."
+        )
 
     problem_fn = model.to_file(problem_fn)
 
@@ -1117,7 +1136,7 @@ def run_mindopt(
         try:
             dual = pd.Series({c.constrname: c.DualSoln for c in m.getConstrs()})
             dual = set_int_index(dual)
-        except mindoptpy.MindoptError:
+        except (mindoptpy.MindoptError, AttributeError):
             logger.warning("Dual values of MILP couldn't be parsed")
             dual = pd.Series(dtype=float)